requirements.txt
scenedetect
opencv-python
requests>=2.25.1
python-dotenv>=0.15.0
pathlib>=1.0.1